Move fastapi example into separate samples folder
